Lots of changes
The codebase is still in a flux, and I haven't been committing my
changes my discipline. I won't bother picking this changeset apart so I
applogize to any source code archeologist.

This commit does represent a state of:

 * transport protocols supported: ssh
 * git protocol supported: v2
 * record based definition of git objects
 * ls-refs working
 * fetch working, partially
    * can clone repos
    * can fetch repos with no change needed
    * can fetch repos with new objects

When I say "working" is that simple situations have been seen to work.

Git objects are stored in memory (a map, in srcd_repo). Have a look at
srcd_sup for how to define repos and objects.

In its current state, since we don't support push operations, the only
way to define new objects it to carefully handcraft them - or, easier:
generate them using git and carefully copy the values (including any
trailing newlines of blob data).

To inspect the content of a compressed object, I've found

    pigz -zd <objects/ff/ffffff....

to be useful, since it can directly work with deflate compressed files.

To be able to pass environments over ssh, I had to hack the erlang ssh
app to be able to pass the collected envs to the ssh exec fun. This is
the only way the ssh client will flag that it supports v2. (Perhaps I
can force a compatible client to use v2 by assuming it to be true? Not
tested and would probably be unwise, since we do want to support v0
clients at some point.)