Rework rebar/relx config to support running integration tests
We have to define a relase to have something to build, so that we have
something to run so that we have something to test :). That's all this
is, for now. For my own needs, I see myself embedding this as an OTP
application within my own (unrelated) release, but I think letting srcd
define its own release for standalone operations seems likely. But
that's later.