srcd_ssh: Resolve listen hostnames
We want to support interface selection by configuring sshd to listen on
a specific hostname (e.g. localhost); but the ssh:daemon doesn't seem to
accept this (even though the documentation arguably suggests that it
should be). We resolve it ourself, and in turn loose the ability to
configure with atom values like "any" or "loopback".

We will have to revisit this.