tests: Try to use a (hard coded) random loopback address
We get EADDRNOTAVAIL errors from the ssh client when trying to connect
to the daemon. I think this is because 1) the github action docker
engine is configured to share network stack between containers (?), and
2) the random client source port range is exhausted.

We can try to use another ip address within the vast 127/8 network and
see if it makes any difference.