tests: wip: make tests run smoother over various git versions
Github actions (well, the docker image i use) only has git v2.20.2;
supporting older gits would be useful for real world reasons, so I'm
making an effort to solve the issues I see rather than trying to have a
later git during testing.

We can reproduce some of the errors we see if we force later versions of
git to use protocol=v0.

This is a wip commit that leaves things in worse state than it was, but
at least I reproduce some of the errors locally with this. Pushing to
master anyways, because I'm an agent of chaos.