tests: automatically figure some things out based on naming
we can lock in some kind of naming convention for the tests, so that the
test framework can figure some basic things out. specifically, for
protocol tests, one of the most fundamental question is: what protocol
version? if we name all the tests like "NN-proto-vX-command.t", tap.sh
can extract the protocol version and set it for us.