zlib: Wrap erlang's zlib module for progress control
We need to know how many compressed bytes zlib consumes. We do this by
wrapping the IO so that we can feed it one byte at a time, under some
arbitrary conditions, we try to complete the stream. If it suceeds, we
should be good, otherwise we restart the stream (with the progress so
far fed directly). At the end of the packfile processing, we will verify
a hash signature of the packfile as a whole, so if we did mess up
somehow, things would fail at this point if not before.

This seems to me both unrobust and a huge performance bottleneck. One
mitigating factor is that we read not against real IO but against an
internal erlang buffer. Still, there is a lot overhead for for every
zlib byte we process (especially if we try and fail to end the stream).
Improving on this would probably mean reimplementing (or forking
erlang's) zlib. But for now, it allows us to at least process the
packiles and pass our tests.