ssh: Don't interpret ^C (NB: unsuitable for upstreaming!)
As the git pack protocol is a binary protocol, we can't assume anything
about specific byte values. We cannot interpret control characters. When
git says a packfile contains three objects, it does mean that erlang
should kill the stream.

This change just removes all ^C processing, which works nicely in my
specific case, it seems. This change is not suitable for upstreaming, as
it changes existing behavior. But it does give a hint that the ^C
processing may belong somewhere else, or at least be configurable.
(Maybe I just missed something?)