srcd_object: Better support for author/committer lines
We now actually extract the committer and author lines from the object,
instead of faking a hard coded one. We still don't parse the author
line, but cheat a bit by passing it along as is (as the name!). This is
enough to be able to reproduce the commit object so that the digest
matches, but will require further improvements.