srcd_sqlite3: module to dump repo to disk as an sqlite db
This is used as a possible strategy for persistence; i.e. dump all
git objects to disk so that we can restart without losing data otherwise
only kept in memory.

Arguments for sqlite: if anything goes wrong, we can extract the sqlite
and either debug it or load it with other means. The format and tools to
deal with the databases are ubiqutous and well understood.

Arguments against: esqlite3 seems to not be super stable/mature/bugfree.
It gives undocumented return values on success, there's been a few
segfault bugs in the passed. The C nature of the sqlite3 bindings expose
the users of esqlite3 to various behaviors of Beam that seem quite
complicated to debug.

I'll commit the code, but note that this is not something I recommend at
this point.

I did not add esqlite as a dependency, nor did i add it to srcd.app.src.
If you want to use sqlite, you have to do that yourself.