git-clone-unpacked: improve robustness
Now we clone the repo, not only of the packed objects, but also of loose
objects (we can just copy them). We also don't fail because of a lack of
packs.

We enable --quiet for git operations, since the output is of no use when
run from the test suite.