Restructure test infrastructure, add support for fixtures
Less important what version of the git protocol we want to use, more
important what kind of environment we are setting up for the tests.

The server is now started once per suite, instead of once per test run.
This is to allow for test suites specializing in write operations, that
may otherwise affect other tests.

The badly named tap.sh is now in charge for more of the environment set
up, and run.sh takes care of any set up that still is for whole test run
(like ssh configuration).