Add test suite "mutations" (WIP)
This test suite sets up specific repos for its use, and preform changes
to them (pushing new commits).

The commented out portions of the test suite currently fails on a git
push operation; "all" I have to go on is this:

  SSH Error: exception {case_clause,0}

no further context :). Somewhere - possibly in my code - there is a case
that does not expect to get a 0. My best guess is that the zlib inflator
got confused on where to stop reading (probably reading one byte less
than what it should). For this reason, I improved the unit tests for
srcd_zlib, but wasn't able to reproduce the issue. My zlib wrapper is
very wonky, so I still suspect it to be the culprit somehow, but I've
suspected it before only to find trivial bugs elsewhere to be the cause
(e.g. in the tests themselves, here my hoping :)).