Add new app: flate, zlib implementation (stub)
Right now, iff you happen to want to deflate something that is the result
of inflating the empty string, you are in luck!

This is my attempt to make "my current zlib implementation" (which is
just a weird wrapper around OTP's zlib implementation) less wonky. For
the purposes of gitting, we need insights into the inner workings of the
deflate algorithm (or rather, the inflate algorithm). OTP's zlib doesn't
provide enough to satisfy our needs. (We still just wrap their deflate
though.)

Unit tests are failing because, again, this is only a stub.