Add profile attribute for repos
This will serve to select which behavior repo will adapt; regular repos
accepts new changes like normal, but some profiles may put requirements
on the format or metadata of the change - ranging from "commit is pgp
signed" to "passes test suite in CI" or "changes successfully deployed".

For instance, a "dns" profile may be implemented to only accept changes
that only touches files according to a certain pattern and adhere to a
certain format representing the dns zone.