ssh: Crash harder on exceptions
We lose all diagnostics except the message with how the ssh cli handled
errors. This change disables exception catching and lets the exceptions
(potentially) crash the application instead.

As an example: how to you troubleshoot ssh connections closing with only
the error "undef" to go on - on the client side? Besides being hard to
troubleshoot, we don't want to send application internal system errors
to the client. We would be happy to say "internal server error" or
something.