flate_lz77: add lazy_decode and resolve
lazy_decode will just decode the lz77 instruction, but won't execute it.
resolve will take the instruction and execute it (without decoding
anything from the input stream). the decode function from prior is now
simply the composition of these two functions.

the idea is that we can use lazy_decode if we in the client code want
to defer the resolving for a later step.