flate: Do lazy lz77 decoding
A little bit lazy, this is only preparing to fix the bug, but we need to
find a good place to do this processing. Just at this point where i
placed the lz77 resolving, we also convert the list to binary. We should
probably consider defering that to later as well.